+++ src/main/kotlin/dev/cursedmc/libreg/registry/Registry.kt
@@ -1,0 +1,13 @@
+@file:JvmName("RegistryUtil")
+
+package dev.cursedmc.libreg.registry
+
+import net.minecraft.util.Identifier
+
+var namespace: String = "minecraft"
+
+fun init(ns: String) {
+	namespace = ns
+}
+
+fun ident(id: String): Identifier = Identifier(namespace, id)