+++ modrinth/CHANGELOG.md
@@ -1,3 +1,1 @@
- - Major bug fixes
-   - Server no longer crashes when loading (due to NoClassDefFoundError)
- - Removed RenderLayer from overloaded arguments in `block`, `blockItem`, and `blockEntity`.
+ - Update package name from `dev.tehc.libreg` to `dev.cursedmc.libreg`